payments/src/routes/__test__/new.test.ts
import { OrderStatus } from '@bhuone/common';
import mongoose from 'mongoose';
import request from 'supertest';
import { app } from '../../app';
import { Order } from '../../models/order';

it('should return 404 when purchase an order that does not exists', async () => {
  await request(app)
    .post('/api/payments')
    .set('Cookie', global.signin())
    .send({
      stripToken: 'sdfs',
      orderId: mongoose.Types.ObjectId().toHexString()
    })
    .expect(404);
});

it('should return 401 if order does not belong to the user', async () => {

  const order = Order.build({
    id: mongoose.Types.ObjectId().toHexString(),
    userId: mongoose.Types.ObjectId().toHexString(),
    version: 0,
    status: OrderStatus.Created,
    price: 20,
  });

  await order.save();

  await request(app)
    .post('/api/payments')
    .set('Cookie', global.signin())
    .send({
      orderId: order.id,
      stripToken: 'sdfs',
    })
    .expect(401);
});

it('should return 400 when attempting to create payment for an order that is already cancelled', async () => {
  const userId = mongoose.Types.ObjectId().toHexString();

  const order = Order.build({
    id: mongoose.Types.ObjectId().toHexString(),
    userId,
    version: 0,
    status: OrderStatus.Cancelled,
    price: 20,
  });

  await order.save();

  await request(app)
    .post('/api/payments')
    .set('Cookie', global.signin(userId))
    .send({
      orderId: order.id,
      stripToken: 'sdfs',
    })
    .expect(400);

});
